+++ examples/classification/plot_sadinle2019_example.py
@@ -14,16 +14,14 @@ We model the data with Gaussian Naive Bayes classifier
 
 Prediction sets are estimated by :class:`mapie.classification.MapieClassifier`
 from the distribution of the softmax scores of the true labels for three
-alpha values (0.2, 0.1, and 0.05) giving different class coverages.
+alpha values (0.2, 0.1, and 0.05) giving different class coverage levels.
 
-When the class coverage is not large enough, the prediction sets can be empty
-when the model is uncertain at the border between two labels. The null region
-disappears for larger class coverages but ambiguous classification regions
-arise with several labels included in the prediction sets.
+When the class coverage is not large enough, the prediction sets can be empty.
+This happens because the model is uncertain at the border between two labels.
+These so-called null regions disappear for larger coverage levels.
 """
 import numpy as np
 import matplotlib.pyplot as plt
-
 from sklearn.naive_bayes import GaussianNB
 
 from mapie.classification import MapieClassifier
@@ -31,27 +29,34 @@ from mapie.classification import MapieClassifier
 
 # Create training set from multivariate normal distribution
 centers = [(0, 3.5), (-2, 0), (2, 0)]
-covs = [[[1, 0], [0, 1]], [[2, 0], [0, 2]], [[5, 0], [0, 1]]]
-x_min, x_max, y_min, y_max, step = -5, 7, -5, 7, 0.1
+# covs = [[[1, 0], [0, 1]], [[2, 0], [0, 2]], [[5, 0], [0, 1]]]
+covs = [np.eye(2), np.eye(2)*2, np.diag([5, 1])]
+x_min, x_max, y_min, y_max, step = -6, 8, -6, 8, 0.1
 n_samples = 500
-alphas = [0.2, 0.1, 0.05]
-X_train = np.zeros((3*n_samples, 2))
-i = 0
-for center, cov in zip(centers, covs):
-    (
-        X_train[i*n_samples:(i+1)*n_samples, 0],
-        X_train[i*n_samples:(i+1)*n_samples, 1]
-    ) = np.random.multivariate_normal(center, cov, n_samples).T
-    i += 1
-y_train = np.stack([i for i in range(3) for _ in range(n_samples)]).ravel()
+n_classes = 3
+alpha = [0.2, 0.1, 0.05]
+np.random.seed(42)
+X_train = np.vstack([
+    np.random.multivariate_normal(center, cov, n_samples)
+    for center, cov in zip(centers, covs)
+])
+y_train = np.hstack([np.full(n_samples, i) for i in range(n_classes)])
+# y_train = np.stack([i for i in range(3) for _ in range(n_samples)]).ravel()
+
 
 # Create test from (x, y) coordinates
-X_test = np.stack([
-    [x, y]
-    for x in np.arange(x_min, x_max, step)
-    for y in np.arange(x_min, x_max, step)
-])
+# X_test = np.stack([
+#     [x, y]
+#     for x in np.arange(x_min, x_max, step)
+#     for y in np.arange(x_min, x_max, step)
+# ])
+xx, yy = np.meshgrid(
+    np.arange(x_min, x_max, step), np.arange(x_min, x_max, step)
+)
+X_test = np.stack([xx.ravel(), yy.ravel()], axis=1)
 
+# print(xx.shape, yy.shape, X_test.shape)
+# print(np.stack((xx.ravel(), yy.ravel()), axis=1))
 
 # Apply MapieClassifier on the dataset to get prediction sets
 clf = GaussianNB().fit(X_train, y_train)
@@ -60,19 +65,22 @@ y_pred_proba = clf.predict_proba(X_test)
 y_pred_proba_max = np.max(y_pred_proba, axis=1)
 mapie = MapieClassifier(estimator=clf, cv="prefit")
 mapie.fit(X_train, y_train)
-y_pred_mapie, y_pi_mapie = mapie.predict(X_test, alpha=alphas)
+y_pred_mapie, y_ps_mapie = mapie.predict(X_test, alpha=alpha)
 
 # Plot the results
-tab10 = plt.cm.get_cmap('Purples', 4)
-colors = ['#1f77b4', '#ff7f0e', '#2ca02c', '#d62728']
-y_pred_col = [colors[int(i)] for _, i in enumerate(y_pred_mapie)]
+tab10 = plt.cm.get_cmap("Purples", 4)
+colors = {0: "#1f77b4", 1: "#ff7f0e", 2:  "#2ca02c", 3: "#d62728"}
+y_pred_col = list(map(colors.get, y_pred_mapie))
+y_train_col = list(map(colors.get, y_train))
+# colors = ["#1f77b4", "#ff7f0e", "#2ca02c", "#d62728"]
+# y_pred_col = [colors[int(i)] for _, i in enumerate(y_pred_mapie)]
 y_train_col = [colors[int(i)] for _, i in enumerate(y_train)]
 fig, axs = plt.subplots(1, 4, figsize=(20, 4))
 axs[0].scatter(
     X_test[:, 0],
     X_test[:, 1],
     color=y_pred_col,
-    marker='.',
+    marker=".",
     s=10,
     alpha=0.4
 )
@@ -80,18 +88,18 @@ axs[0].scatter(
     X_train[:, 0],
     X_train[:, 1],
     color=y_train_col,
-    marker='o',
+    marker="o",
     s=10,
-    edgecolor='k'
+    edgecolor="k"
 )
 axs[0].set_title("Predicted labels")
-for i, alpha in enumerate(alphas):
-    y_pi_sums = y_pi_mapie[:, :, i].sum(axis=1)
+for i, alpha_ in enumerate(alpha):
+    y_ps_sums = y_ps_mapie[:, :, i].sum(axis=1)
     num_labels = axs[i+1].scatter(
         X_test[:, 0],
         X_test[:, 1],
-        c=y_pi_sums,
-        marker='.',
+        c=y_ps_sums,
+        marker=".",
         s=10,
         alpha=1,
         cmap=tab10,
@@ -99,5 +107,5 @@ for i, alpha in enumerate(alphas):
         vmax=3
     )
     cbar = plt.colorbar(num_labels, ax=axs[i+1])
-    axs[i+1].set_title(f"Number of labels for alpha={alpha}")
+    axs[i+1].set_title(f"Number of labels for alpha={alpha_}")
 plt.show()