+++ HISTORY.rst
@@ -6,7 +6,7 @@ History
 ------------------
 
 * Renaming estimators.py module to regression.py
-* New classification.py module with MapieClassifier class, that estimate prediction sets from softmax score
+* New classification.py module with MapieClassifier class, that estimates prediction sets from softmax score
 * New set of unit tests for classification.py module
 * Modification of the documentation architecture
 * Split example gallery into separate regression and classification galleries